feat: token fix
