+++ __tests__/end-to-end/03customer_products.test.js
@@ -107,7 +107,7 @@ describe(requirement(13), () => {
   })
 });
 
-describe.only(requirement(14), () => {
+describe(requirement(14), () => {
   const cards = products.state01;
 
   test("O avaliador testará se os dados de cada card condizem com os dados esperados",
