+++ back-end/jwt.evaluation.key
@@ -1,1 +1,1 @@
-secret_key
+M1nAwaTT2hEkN41QnmCkqfyzbohr4snE
