+++ back-end/src/api/services/login.js
@@ -3,8 +3,11 @@ const { User } = require('../../database/models');
 
 const login = async ({ email, password }) => {
   const hash = md5(password);
-  const result = await User.findOne({ where: { email, password: hash } });
-  return result; 
+  const result = await User.findOne(
+    { where: { email, password: hash } },
+    );
+  const { id, name, role } = result;
+  return { id, name, email, role }; 
 };
 
 module.exports = {
